requirements.txt
fastapi
uvicorn
python-multipart
Pillow
aiofiles
folium